requirements.txt
streamlit==0.69.2
numpy==1.18.5
pandas == 1.2.3 
matplotlib == 3.4.1
emoji-country-flag == 1.2.4 
emoji == 1.2.0 
ast == 0.0.1
pywaffle == 0.6.1
statsbombpy == 1.0